refactor: :recycle: move content into `main.py`, use scripts as functions (#29)
# Description

Something has been bugging me for a while with our potential workflow,
especially around the resources. One idea could be that we use `main.py`
in the parent folder to coordinate all the tasks, and keep the `scripts`
to hold the python functions that are used in `main.py`. Then we can
simplify to `write_properties()`, `read_properties()` and no updating
function. And we use the `PackageProperties` in `main.py` as the source
of truth for the properties.

Then we can make templates to create a script of the extracted resource
properties as functions.

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.